pyproject.toml
[project]
name = "openai-sdk-restaurant-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "openai-agents[voice]>=0.2.8",
]
